requirements.txt
pandas
scikit-learn
matplotlib
opencv-python
pillow
torch
torchvision